+++ packages/components/avatar/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sl-design-system/avatar
 
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies [[`3295fa8`](https://github.com/sl-design-system/components/commit/3295fa8a92a7b0284a422232884f5fef77aa8537), [`3295fa8`](https://github.com/sl-design-system/components/commit/3295fa8a92a7b0284a422232884f5fef77aa8537)]:
+  - @sl-design-system/shared@0.2.10
+  - @sl-design-system/tooltip@0.0.25
+
 ## 0.2.6
 
 ### Patch Changes